Re-arranging message enums for broker move
