+++ cli/signal/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-types = {path="../../fltypes"}
-raw = {path="../../modules"}
+flutils = {path="../../flutils"}
+flmodules = {path="../../flmodules"}
 common = {path = "../../common"}
 serde = "1.0.127"
 serde_json = "1.0.66"
