+++ common/Cargo.toml
@@ -8,8 +8,8 @@ version = "0.1.0"
 
 [dependencies]
 transitive_from = "0.1.3"
-types = {path = "../fltypes"}
-raw = {path = "../modules"}
+flutils = {path = "../flutils"}
+flmodules = {path = "../flmodules"}
 async-trait = "0.1.51"
 rand = "0.7"
 serde = { version = "1.0.127", features = ["derive"] }
