+++ common/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod broker;
 pub mod node;
 pub mod signal;
-pub mod broker_join;
