+++ common/src/node/stats.rs
@@ -1,10 +1,3 @@
-use crate::{
-    broker::{Broker, Subsystem},
-    node::{
-        modules::messages::NodeMessage,
-        network::{BrokerNetwork, NetworkConnectionState},
-    },
-};
 use core::f64;
 use std::{
     collections::HashMap,
@@ -12,19 +5,21 @@ use std::{
 };
 use thiserror::Error;
 
+use flutils::{nodeids::U256, utils::now};
+
 use crate::{
-    broker::{BrokerError, BrokerMessage, SubsystemListener},
+    broker::{Broker, Subsystem},
+    broker::{BrokerError, SubsystemListener},
     node::{
         config::{NodeConfig, NodeInfo},
-        modules::messages::{Message, MessageV1},
-        network::{connection_state::CSEnum, ConnStats},
+        modules::messages::{BrokerMessage, Message, MessageV1, NodeMessage},
+        network::{connection_state::CSEnum, BrokerNetwork, ConnStats, NetworkConnectionState},
         node_data::NodeData,
         timer::BrokerTimer,
         version::VERSION_STRING,
     },
     signal::web_rtc::{ConnType, NodeStat, WebRTCConnectionState},
 };
-use types::{nodeids::U256, utils::now};
 
 #[derive(Debug, Error)]
 pub enum SNError {
