--- fltypes/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "types"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0.30"
-rand = "0.8.4"
-serde = "1.0.136"
-sha2 = "0.10.1"
-futures = "0.3.19"
-chrono = "0.4.19"
