--- fltypes/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod data_storage;
-pub mod nodeids;
-pub mod utils;
