--- modules/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "raw"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-types = { path = "../fltypes" }
-serde = { version = "1.0.127", features = ["derive"] }
-serde_derive = "1.0.127"
-serde_json = "1.0.66"
-serde_yaml = "0.8"
-toml = "0.5.8"
-sha2 = "0.9.5"
-rand = "0.7"
-thiserror = "1.0.26"
-tracing = "0.1.29"
-log = "0.4.14"
-itertools = "0.10.3"
