--- modules/README.md
@@ -1,7 +1,0 @@
-# Raw Fledger modules
-
-This crate holds the raw modules for Fledger.
-Every module is as independant as possible, only depending on the
-`NodeID` and `NodeIDs` structures.
-
-Eventually wrappers for other crates might also show up in here.
