--- modules/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod random_connections;
-pub mod gossip_events;
