+++ wasm-webrtc/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 node = []
 
 [dependencies]
-types = {path = "../fltypes"}
+flutils = {path = "../flutils"}
 common = {path = "../common"}
 wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"] }
 js-sys = "0.3.51"
