+++ wasm-webrtc/tests/webtests/test_webrtc.rs
@@ -16,7 +16,7 @@ use common::{
         websocket::{MessageCallback, WSError, WSMessage, WebSocketConnection},
     },
 };
-use types::{
+use flutils::{
     nodeids::U256, data_storage::TempDSB,
 };
